Category Fixes
Correct erroneously named categories and incorrect category code (caused by a merge error?).

See #1417
