+++ app/src/frontend/building/data-containers/age-history.tsx
@@ -1,801 +1,554 @@
 import React, { Fragment } from 'react';
 
-import { buildingUserFields, commonSourceTypes, dataFields } from '../../config/data-fields-config';
+import { commonSourceTypes, dataFields } from '../../config/data-fields-config';
 import DataEntry from '../data-components/data-entry';
 import SelectDataEntry from '../data-components/select-data-entry';
 import withCopyEdit from '../data-container';
 import Verification from '../data-components/verification';
-
 import { CategoryViewProps } from './category-view-props';
 import InfoBox from '../../components/info-box';
 import { DataEntryGroup } from '../data-components/data-entry-group';
 import { MultiDataEntry } from '../data-components/multi-data-entry/multi-data-entry';
 import { LogicalDataEntry } from '../data-components/logical-data-entry/logical-data-entry';
-import { Category } from '../../config/categories-config';
-import PlanningDataOfficialDataEntry from '../data-components/planning-data-entry';
 import NumericDataEntry from '../data-components/numeric-data-entry';
 import { useDisplayPreferences } from '../../displayPreferences-context';
-import UserOpinionEntry from '../data-components/user-opinion-data-entry';
-import { Link } from 'react-router-dom';
-import NumericDataEntryWithFormattedLink from '../data-components/numeric-data-entry-with-formatted-link';
-
-const currentTimestamp = new Date().valueOf();
-const milisecondsInYear = 1000 * 60 * 60 * 24 * 365;
-
-// there is already "parseDate" in helpers
-// but it is using timestamp as input, while this one
-// uses lower accuracy (as actual data is using the same accuracy)
-function parseDateSpecifiedWithDailyAccuracy(isoUtcDate: string): Date {
-    const [year, month, day] = isoUtcDate.match(/^(\d{4})-(\d\d)-(\d\d)$/)
-        .splice(1)
-        .map(x => parseInt(x, 10));
-    return new Date(Date.UTC(year, month-1, day));
-}
-
-function isArchived(item) {
-    const decisionDate = item.decision_date;
-    if(decisionDate != null) {
-        if ((currentTimestamp - parseDateSpecifiedWithDailyAccuracy(decisionDate).valueOf()) > milisecondsInYear) {
-            return true;
-        }
-    }
-    if(item.registered_with_local_authority_date != null) {
-        if ((currentTimestamp - parseDateSpecifiedWithDailyAccuracy(item.registered_with_local_authority_date).valueOf()) > milisecondsInYear) {
-            return true;
-        }
-    }
-    return false;
-}
+import YearDataEntry from '../data-components/year-data-entry';
+import { DynamicsBuildingPane, DynamicsDataEntry } from './dynamics/dynamics-data-entry';
+import { FieldRow } from '../data-components/field-row';
 
 /**
 * Age & History view/edit section
 */
 const AgeHistoryView: React.FunctionComponent<CategoryViewProps> = (props) => {
-    const queryParameters = new URLSearchParams(window.location.search);
-    const subcat = queryParameters.get("sc");
-    
-    const switchToExpectedApplicationMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('community_expected_planning_application_total')
-    }
-    const switchToBuildingProtectionMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('planning_combined')
-    }
-    const switchToAllPlanningApplicationsMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('planning_applications_status_all')
-    }
-    const switchToLastTwelveMonthsMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('planning_applications_status_recent')
-    }
-    const switchToLastThirtyDaysMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('planning_applications_status_very_recent')
-    }
-    const switchToLocalSignificanceMapStyle = (e) => {
-        e.preventDefault();
-        props.onMapColourScale('community_local_significance_total')
-    }
-
-    const { housing, housingSwitchOnClick, creative, creativeSwitchOnClick, vista, vistaSwitchOnClick, parcel, parcelSwitchOnClick, conservation, conservationSwitchOnClick, darkLightTheme } = useDisplayPreferences();
-    
-    const communityLinkUrl = `/${props.mode}/${Category.Community}/${props.building.building_id}`;
     const currentYear = new Date().getFullYear();
 
+    const building = props.building;
+    const thisYear = (new Date()).getFullYear();
+    const currentBuildingConstructionYear = building.date_year || undefined;
+
+    const { historicData, historicDataSwitchOnClick, darkLightTheme } = useDisplayPreferences();
+    const { historicMap, historicMapSwitchOnClick } = useDisplayPreferences();
+
+    const switchToSurvivalMapStyle = (e) => {
+        e.preventDefault();
+        props.onMapColourScale('survival_status');
+        historicMapSwitchOnClick(e);
+        
+        if (historicData === 'enabled') {
+           historicDataSwitchOnClick(e);
+        }
+    }
+
+    const switchToSurvivalDataStyle = (e) => {
+        e.preventDefault();
+        props.onMapColourScale('survival_status');
+        historicDataSwitchOnClick(e);
+
+        if (historicMap === 'enabled') {
+            historicMapSwitchOnClick(e);
+        }
+    }
+
+    const switchToAgeMapStyle = (e) => {
+        e.preventDefault();
+
+        if (historicData === 'enabled') {
+            historicDataSwitchOnClick(e);
+        }
+        if (historicMap === 'enabled') {
+            historicMapSwitchOnClick(e);
+        }
+
+        props.onMapColourScale('date_year');
+    }
+
+    const switchToStylePeriodMapStyle = (e) => {
+        e.preventDefault();
+        props.onMapColourScale('typology_style_period')
+    }
+
+    const queryParameters = new URLSearchParams(window.location.search);
+    const subcat = queryParameters.get("sc");
+
     return (
-         <Fragment>
-            <DataEntryGroup name="Planning" collapsed={subcat==null || (subcat!="1" && subcat!="2" && subcat!="3")}>
-                <DataEntryGroup name="Current planning applications" collapsed={subcat==null || (subcat!="1" && subcat!="2" && subcat!="3")}>
-                    <DataEntryGroup name="Official data" collapsed={subcat==null || subcat!="1"}>
-                        <InfoBox>
-                            This section provides data on active applications. We define these as applications with any activity in the last year.
-                            <br />
-                            To comment on an application follow the application link if provided, or visit the relevant local authority's planning page.
-                        </InfoBox>
-                        {props.mapColourScale != "planning_applications_status_all" ?
-                            <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark key-button sub-subcategory-button`} onClick={switchToAllPlanningApplicationsMapStyle}>
-                                {'Click to view official planning application data'}
-                            </button>
-                            :
-                            <></>
-                        }
-                        {props.building.planning_data ?
-                            <PlanningDataOfficialDataEntry  
-                                shownData={props.building.planning_data.filter(item => isArchived(item) == false)}
-                                messageOnMissingData={
-                                    props.building.planning_data.length > 0 ?
-                                        "Only past application data is currently available for this site"
-                                        :
-                                        "No live planning data are currently available for this building from the Planning London Datahub."
-                                }
-                            />
-                        : <></>
-                        }
-                    </DataEntryGroup>
-                    <DataEntryGroup name="Year of completion" collapsed={subcat==null || subcat!="2"}>
-                        <LogicalDataEntry
-                            slug='planning_crowdsourced_site_completion_status'
-                            title={dataFields.planning_crowdsourced_site_completion_status.title}
-                            tooltip={dataFields.planning_crowdsourced_site_completion_status.tooltip}
-                            value={props.building.planning_crowdsourced_site_completion_status}
+        <Fragment>
+            <DataEntryGroup name="Architectural style/historial period" collapsed={subcat==null || subcat!="2"}>
+                {(props.mapColourScale != "typology_style_period") ? 
+                    <button className={`map-switcher-inline enabled-state btn btn-outline btn-outline-dark key-button`} onClick={switchToStylePeriodMapStyle}>
+                        Click to show architectural style.
+                    </button>
+                :
+                    <></>
+                }
+                <SelectDataEntry
+                    title={dataFields.typology_style_period.title}
+                    slug="typology_style_period"
+                    value={props.building.typology_style_period}
+                    tooltip={dataFields.typology_style_period.tooltip}
+                    options={dataFields.typology_style_period.items}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                />
+                <Verification
+                    slug="typology_style_period"
+                    allow_verify={props.user !== undefined && props.building.typology_style_period !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("typology_style_period")}
+                    user_verified_as={props.user_verified.typology_style_period}
+                    verified_count={props.building.verified.typology_style_period}
+                />
+                <SelectDataEntry
+                    title={dataFields.typology_style_period_source_type.title}
+                    slug="typology_style_period_source_type"
+                    value={props.building.typology_style_period_source_type}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.typology_style_period_source_type.tooltip}
+                    placeholder={dataFields.typology_style_period_source_type.example}
+                    options={dataFields.typology_style_period_source_type.items}
+                    />
+                {(props.building.typology_style_period_source_type == commonSourceTypes[0] ||
+                    props.building.typology_style_period_source_type == commonSourceTypes[1] ||
+                    props.building.typology_style_period_source_type == null) ? <></> :
+                    <>
+                        <MultiDataEntry
+                            title={dataFields.typology_style_period_source_links.title}
+                            slug="typology_style_period_source_links"
+                            value={props.building.typology_style_period_source_links}
+                            mode={props.mode}
                             copy={props.copy}
                             onChange={props.onChange}
+                            tooltip={dataFields.typology_style_period_source_links.tooltip}
+                            placeholder="https://..."
+                            editableEntries={true}
+                            isUrl={true}
+                        />
+                    </>
+                }
+            </DataEntryGroup>
+            <DataEntryGroup name="Building age/construction date" collapsed={subcat==null || subcat!="1"}>
+                {(props.mapColourScale != "date_year") ? 
+                        <button className={`map-switcher-inline enabled-state btn btn-outline btn-outline-dark key-button`} onClick={switchToAgeMapStyle}>
+                            Click to show building age.
+                        </button>
+                :
+                    <></>
+                }
+                <YearDataEntry
+                    year={props.building.date_year}
+                    upper={props.building.date_upper}
+                    lower={props.building.date_lower}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+
+                    allow_verify={props.user !== undefined && props.building.date_year !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("date_year")}
+                    user_verified_as={props.user_verified.date_year}
+                    verified_count={props.building.verified.date_year}
+                    
+                    allow_verify_upper={props.user !== undefined && props.building.date_upper !== null && !props.edited}
+                    onVerify_upper={props.onVerify}
+                    user_verified_upper={props.user_verified.hasOwnProperty("date_upper")}
+                    user_verified_as_upper={props.user_verified.date_upper}
+                    verified_count_upper={props.building.verified.date_upper}
+                    
+                    allow_verify_lower={props.user !== undefined && props.building.date_lower !== null && !props.edited}
+                    onVerify_lower={props.onVerify}
+                    user_verified_lower={props.user_verified.hasOwnProperty("date_lower")}
+                    user_verified_as_lower={props.user_verified.date_lower}
+                    verified_count_lower={props.building.verified.date_lower}
+                    />
+                <NumericDataEntry
+                    title={dataFields.facade_year.title}
+                    slug="facade_year"
+                    value={props.building.facade_year}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    step={1}
+                    min={1}
+                    max={currentYear}
+                    tooltip={dataFields.facade_year.tooltip}
+                    />
+                <Verification
+                    slug="facade_year"
+                    allow_verify={props.user !== undefined && props.building.facade_year !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("facade_year")}
+                    user_verified_as={props.user_verified.facade_year}
+                    verified_count={props.building.verified.facade_year}
+                    />
+                <hr/>
+                <SelectDataEntry
+                    title={dataFields.date_source_type.title}
+                    slug="date_source_type"
+                    value={props.building.date_source_type}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.date_source_type.tooltip}
+                    options={dataFields.date_source_type.items}
+                    placeholder={dataFields.date_source_type.example}
+                    />
+                {(props.building.date_source_type == dataFields.date_source_type.items[0] ||
+                    props.building.date_source_type == dataFields.date_source_type.items[1] ||
+                    props.building.date_source_type == null) ? <></> :
+                    <>
+                        <MultiDataEntry
+                            title={dataFields.date_source_links.title}
+                            slug="date_source_links"
+                            value={props.building.date_source_links}
                             mode={props.mode}
+                            copy={props.copy}
+                            onChange={props.onChange}
+                            tooltip={dataFields.date_source_links.tooltip}
+                            placeholder="https://..."
+                            editableEntries={true}
+                            isUrl={true}
                         />
-                        <Verification
-                            slug="planning_crowdsourced_site_completion_status"
-                            allow_verify={props.user !== undefined && props.building.planning_crowdsourced_site_completion_status !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_crowdsourced_site_completion_status")}
-                            user_verified_as={props.user_verified.planning_crowdsourced_site_completion_status}
-                            verified_count={props.building.verified.planning_crowdsourced_site_completion_status}
+                    </>
+                }
+            </DataEntryGroup>
+            <DataEntryGroup name="Cladding, extensions and retrofits" collapsed={subcat==null || subcat!="3"}>
+                <NumericDataEntry
+                    slug='age_cladding_date'
+                    title={dataFields.age_cladding_date.title}
+                    value={props.building.age_cladding_date}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    step={1}
+                    min={1}
+                    max={currentYear}
+                    tooltip={dataFields.extension_year.tooltip}
+                    />
+                <Verification
+                    slug="age_cladding_date"
+                    allow_verify={props.user !== undefined && props.building.age_cladding_date !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("age_cladding_date")}
+                    user_verified_as={props.user_verified.age_cladding_date}
+                    verified_count={props.building.verified.age_cladding_date}
+                    />
+                <SelectDataEntry
+                    title={dataFields.age_cladding_date_source_type.title}
+                    slug="age_cladding_date_source_type"
+                    value={props.building.age_cladding_date_source_type}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.age_cladding_date_source_type.tooltip}
+                    options={dataFields.age_cladding_date_source_type.items}
+                    placeholder={dataFields.age_cladding_date_source_type.example}
+                    />
+                {(props.building.age_cladding_date_source_type == dataFields.age_cladding_date_source_type.items[0] ||
+                    props.building.age_cladding_date_source_type == dataFields.age_cladding_date_source_type.items[1] ||
+                    props.building.age_cladding_date_source_type == null) ? <></> :
+                    <>
+                        <MultiDataEntry
+                            title={dataFields.age_cladding_date_source_links.title}
+                            slug="age_cladding_date_source_links"
+                            value={props.building.age_cladding_date_source_links}
+                            mode={props.mode}
+                            copy={props.copy}
+                            onChange={props.onChange}
+                            tooltip={dataFields.age_cladding_date_source_links.tooltip}
+                            placeholder="https://..."
+                            editableEntries={true}
+                            isUrl={true}
                         />
-                        {props.building.planning_crowdsourced_site_completion_status == null ? <></> :
-                            <>
+                    </>
+                }
+                <hr/>
+                <NumericDataEntry
+                    slug='age_extension_date'
+                    title={dataFields.age_extension_date.title}
+                    value={props.building.age_extension_date}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    step={1}
+                    min={1}
+                    max={currentYear}
+                    tooltip={dataFields.extension_year.tooltip}
+                    />
+                <Verification
+                    slug="age_extension_date"
+                    allow_verify={props.user !== undefined && props.building.age_extension_date !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("age_extension_date")}
+                    user_verified_as={props.user_verified.age_extension_date}
+                    verified_count={props.building.verified.age_extension_date}
+                    />
+                <SelectDataEntry
+                    title={dataFields.age_extension_date_source_type.title}
+                    slug="age_extension_date_source_type"
+                    value={props.building.age_extension_date_source_type}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.age_extension_date_source_type.tooltip}
+                    options={dataFields.age_extension_date_source_type.items}
+                    placeholder={dataFields.age_extension_date_source_type.example}
+                    />
+                {(props.building.age_extension_date_source_type == dataFields.age_extension_date_source_type.items[0] ||
+                    props.building.age_extension_date_source_type == dataFields.age_extension_date_source_type.items[1] ||
+                    props.building.age_extension_date_source_type == null) ? <></> :
+                    <>
+                        <MultiDataEntry
+                            title={dataFields.age_extension_date_source_links.title}
+                            slug="age_extension_date_source_links"
+                            value={props.building.age_extension_date_source_links}
+                            mode={props.mode}
+                            copy={props.copy}
+                            onChange={props.onChange}
+                            tooltip={dataFields.age_extension_date_source_links.tooltip}
+                            placeholder="https://..."
+                            editableEntries={true}
+                            isUrl={true}
+                        />
+                    </>
+                }
+                <hr/>
+                <NumericDataEntry
+                    slug='age_retrofit_date'
+                    title={dataFields.age_retrofit_date.title}
+                    value={props.building.age_retrofit_date}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    step={1}
+                    min={1}
+                    max={currentYear}
+                    tooltip={dataFields.extension_year.tooltip}
+                    />
+                <Verification
+                    slug="age_retrofit_date"
+                    allow_verify={props.user !== undefined && props.building.age_retrofit_date !== null && !props.edited}
+                    onVerify={props.onVerify}
+                    user_verified={props.user_verified.hasOwnProperty("age_retrofit_date")}
+                    user_verified_as={props.user_verified.age_retrofit_date}
+                    verified_count={props.building.verified.age_retrofit_date}
+                    />
+                <SelectDataEntry
+                    title={dataFields.age_retrofit_date_source_type.title}
+                    slug="age_retrofit_date_source_type"
+                    value={props.building.age_retrofit_date_source_type}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.age_retrofit_date_source_type.tooltip}
+                    options={dataFields.age_retrofit_date_source_type.items}
+                    placeholder={dataFields.age_retrofit_date_source_type.example}
+                    />
+                {(props.building.age_retrofit_date_source_type == dataFields.age_retrofit_date_source_type.items[0] ||
+                    props.building.age_retrofit_date_source_type == dataFields.age_retrofit_date_source_type.items[1] ||
+                    props.building.age_retrofit_date_source_type == null) ? <></> :
+                    <>
+                        <MultiDataEntry
+                            title={dataFields.age_retrofit_date_source_links.title}
+                            slug="age_retrofit_date_source_links"
+                            value={props.building.age_retrofit_date_source_links}
+                            mode={props.mode}
+                            copy={props.copy}
+                            onChange={props.onChange}
+                            tooltip={dataFields.age_retrofit_date_source_links.tooltip}
+                            placeholder="https://..."
+                            editableEntries={true}
+                            isUrl={true}
+                        />
+                    </>
+                }
+            </DataEntryGroup>
+            <DataEntryGroup name="Lifespan and site history" collapsed={subcat==null || subcat!="4"}>
+                <DataEntryGroup name="Constructions and demolitions on this site" collapsed={subcat==null || subcat!="4"}>
+                    <DynamicsBuildingPane>
+                        <label>Current building (building age data editable above)</label>
+                        <FieldRow>
+                            <div>
                                 <NumericDataEntry
-                                    title={dataFields.planning_crowdsourced_site_completion_year.title}
-                                    slug="planning_crowdsourced_site_completion_year"
-                                    value={props.building.planning_crowdsourced_site_completion_year}
-                                    mode={props.mode}
-                                    copy={props.copy}
-                                    onChange={props.onChange}
-                                    step={1}
-                                    min={1}
-                                    max={currentYear}
-                                    tooltip={dataFields.planning_crowdsourced_site_completion_year.tooltip}
-                                    />
-                                <Verification
-                                    slug="planning_crowdsourced_site_completion_year"
-                                    allow_verify={props.user !== undefined && props.building.planning_crowdsourced_site_completion_year !== null && !props.edited}
-                                    onVerify={props.onVerify}
-                                    user_verified={props.user_verified.hasOwnProperty("planning_crowdsourced_site_completion_year")}
-                                    user_verified_as={props.user_verified.planning_crowdsourced_site_completion_year}
-                                    verified_count={props.building.verified.planning_crowdsourced_site_completion_year}
-                                    />
-                                <SelectDataEntry
-                                    title={dataFields.planning_crowdsourced_site_completion_source_type.title}
-                                    slug="planning_crowdsourced_site_completion_source_type"
-                                    value={props.building.planning_crowdsourced_site_completion_source_type}
-                                    mode={props.mode}
-                                    copy={props.copy}
-                                    onChange={props.onChange}
-                                    tooltip={dataFields.planning_crowdsourced_site_completion_source_type.tooltip}
-                                    options={dataFields.planning_crowdsourced_site_completion_source_type.items}
-                                    placeholder={dataFields.planning_crowdsourced_site_completion_source_type.example}
+                                    slug=''
+                                    title={dataFields.demolished_buildings.items.year_constructed.title}
+                                    value={currentBuildingConstructionYear}
+                                    disabled={true}
+                                    mode='view'
                                 />
-                                {(props.building.planning_crowdsourced_site_completion_source_type == dataFields.planning_crowdsourced_site_completion_source_type.items[0] ||
-                                    props.building.planning_crowdsourced_site_completion_source_type == dataFields.planning_crowdsourced_site_completion_source_type.items[1] ||
-                                    props.building.planning_crowdsourced_site_completion_source_type == null) ? <></> :
+                            </div>
+                            <div>
+                                <NumericDataEntry
+                                    slug=''
+                                    title={dataFields.demolished_buildings.items.year_demolished.title}
+                                    value={undefined}
+                                    placeholder='---'
+                                    disabled={true}
+                                    mode='view'
+                                />
+                            </div>
+                            <div style={{flex: '0 1 27%'}}>
+                                <DataEntry
+                                    slug=''
+                                    title='Lifespan to date'
+                                    value={ (thisYear - currentBuildingConstructionYear) + ''}
+                                    disabled={true}
+                                    mode='view'
+                                />
+                            </div>
+                        </FieldRow>
+                    </DynamicsBuildingPane>
+                    {
+                        currentBuildingConstructionYear == undefined ?
+                            <InfoBox>To add historical records, fill in the building age data (above) first.</InfoBox> :
+                            
+                            <>
+                                <LogicalDataEntry
+                                    slug='dynamics_has_demolished_buildings'
+                                    title={dataFields.dynamics_has_demolished_buildings.title}
+                                    value={building.dynamics_has_demolished_buildings}
+                                    disallowFalse={(building.demolished_buildings?.length ?? 0) > 0}
+                                    disallowNull={(building.demolished_buildings?.length ?? 0) > 0}
+
+                                    onChange={props.onSaveChange}
+                                    mode={props.mode}
+                                    copy={props.copy}
+                                />
+                                {
+                                    building.dynamics_has_demolished_buildings &&
                                     <>
-                                        <MultiDataEntry
-                                            title={dataFields.planning_crowdsourced_site_completion_source_links.title}
-                                            slug="planning_crowdsourced_site_completion_source_links"
-                                            value={props.building.planning_crowdsourced_site_completion_source_links}
-                                            mode={props.mode}
-                                            copy={props.copy}
-                                            onChange={props.onChange}
-                                            tooltip={dataFields.planning_crowdsourced_site_completion_source_links.tooltip}
-                                            placeholder="https://..."
+                                        <DynamicsDataEntry
+                                            
+                                            /* 
+                                                Will clear the edits and new record data upon navigating to another building.
+                                                Should get a better way to do this, plus a way to actually keep unsaved edits.
+                                            */
+                                            key={building.building_id} 
+                                            
+                                            value={building.demolished_buildings}
                                             editableEntries={true}
-                                            isUrl={true}
+                                            slug='demolished_buildings'
+                                            title={dataFields.demolished_buildings.title}
+                                            mode={props.mode}
+                                            onChange={props.onChange}
+                                            onSaveAdd={props.onSaveAdd}
+                                            hasEdits={props.edited}
+                                            maxYear={currentBuildingConstructionYear}
+                                            minYear={50}
                                         />
+                                        {
+                                            props.mode === 'view' &&
+                                                <InfoBox>Switch to edit mode to add/edit past building records</InfoBox>
+                                        }
                                     </>
                                 }
                             </>
-                        }   
-                    </DataEntryGroup>
-                    <DataEntryGroup name="Incomplete/missing data" collapsed={subcat==null || subcat!="3"}>
-                        <LogicalDataEntry
-                            slug='planning_missing_data'
-                            title={dataFields.planning_missing_data.title}
-                            tooltip={dataFields.planning_missing_data.tooltip}
-                            value={props.building.planning_missing_data}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            mode={props.mode}
-                        />
-                        <Verification
-                            slug="planning_missing_data"
-                            allow_verify={props.user !== undefined && props.building.planning_missing_data !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_missing_data")}
-                            user_verified_as={props.user_verified.planning_missing_data}
-                            verified_count={props.building.verified.planning_missing_data}
-                        />
-                        {props.building.planning_missing_data == null ? <></> :
-                            <>
-                                <MultiDataEntry
-                                    title={dataFields.planning_missing_data_links.title}
-                                    slug="planning_missing_data_links"
-                                    value={props.building.planning_missing_data_links}
-                                    mode={props.mode}
-                                    copy={props.copy}
-                                    onChange={props.onChange}
-                                    tooltip={dataFields.planning_missing_data_links.tooltip}
-                                    placeholder="https://..."
-                                    editableEntries={true}
-                                    isUrl={true}
-                                />
-                            </>
-                        }
-                        <InfoBox>
-                            If you feel there are incorrect or missing data relating to this building please contact:  
-                            planningdata@London.gov.uk
-                        </InfoBox>
-                    </DataEntryGroup>
-                </DataEntryGroup>
-                <DataEntryGroup name="Past applications" collapsed={subcat==null || subcat!="4"}>
-                    <InfoBox>
-                        This section provides data on past applications where available from the GLA, including those with no decision in over a year
-                    </InfoBox>
-                    {props.mapColourScale != "planning_applications_status_recent" ?
-                        <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark key-button no-left-margin`} onClick={switchToLastTwelveMonthsMapStyle}>
-                            {'Click to view planning applications from the last 12 months'}
-                        </button>
-                        :
-                        <></>
-                    }
-                    {props.mapColourScale != "planning_applications_status_very_recent" ?
-                        <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark key-button no-left-margin`} onClick={switchToLastThirtyDaysMapStyle}>
-                            {'Click to view planning applications from the last 30 days'}
-                        </button>
-                        :
-                        <></>
-                    }
-                    {props.building.planning_data ?
-                        <PlanningDataOfficialDataEntry  
-                            shownData={props.building.planning_data.filter(item => isArchived(item))}
-                            messageOnMissingData={
-                                props.building.planning_data.length > 0 ?
-                                    "Only current application data is currently available for this site"
-                                    :
-                                    "No live planning data are currently available for this building from the Planning London Datahub."
-                            }
-                        />
-                    : <></>
                     }
                 </DataEntryGroup>
-                <DataEntryGroup name="Possible future applications" collapsed={subcat==null || subcat!="5"}>
-                    <InfoBox type='info'>Click and colour buildings here if you think they may be subject to a future planning application involving demolition. To add your opinion on how well this building works, please also visit the <Link to={communityLinkUrl}>Community</Link> section.</InfoBox>
-                    {props.mapColourScale != "community_expected_planning_application_total" ?
-                        <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark key-button`} onClick={switchToExpectedApplicationMapStyle}>
-                            {'Click to view possible locations of future applications'}
-                        </button>
-                        :
-                        <></>
-                    }
-                    <UserOpinionEntry
-                        slug='community_expected_planning_application'
-                        title={buildingUserFields.community_expected_planning_application.title}
-                        tooltip={buildingUserFields.community_expected_planning_application.tooltip}
-                        userValue={props.building.community_expected_planning_application}
-                        onChange={props.onSaveChange}
-                        mode={props.mode}
-                    />
-                    <InfoBox type='warning'>
-                        Further improvements to this feature are currently being made.
-                    </InfoBox>
-                    <UserOpinionEntry
-                        slug='community_local_significance'
-                        title={buildingUserFields.community_local_significance.title}
-                        tooltip={buildingUserFields.community_local_significance.tooltip}
-                        userValue={props.building.community_local_significance}
-
-                        onChange={props.onSaveChange}
-                        mode={props.mode}
-                    />
-                    {(props.mapColourScale != "community_local_significance_total") ? 
-                        <button className={`map-switcher-inline enabled-state btn btn-outline btn-outline-dark key-button`} onClick={switchToLocalSignificanceMapStyle}>
-                            {'Click to show buildings of local interest.'}
-                        </button>
-                        :
-                        <></>
-                    }
-                </DataEntryGroup>
-                <DataEntryGroup name="Planning zones" collapsed={subcat==null || subcat!="6"}>
-                    <InfoBox>
-                        To view planning zone data for London click the buttons below. You may need to <u>zoom out</u>.
-                        Information on whether a specific building is in a zone will be added automatically in future.
-                    </InfoBox>
-                    <div className={`alert alert-dark`} role="alert" style={{ fontSize: 13, backgroundColor: "#f6f8f9" }}>
-                        <i>
-                            Data in this section comes from <a href="https://www.london.gov.uk/programmes-strategies/planning/digital-planning/planning-london-datahub">the Greater London Authority's Planning London Datahub</a>. Please check the original GLA source when using for planning purposes.
-                            <br />
-                            Specific sources are mentioned in the footer of map for currently enabled layers.
-                        </i>
-                    </div>
-                    <LogicalDataEntry
-                        slug='planning_housing_zone'
-                        title={dataFields.planning_housing_zone.title}
-                        tooltip={dataFields.planning_housing_zone.tooltip}
-                        value={props.building.planning_housing_zone}
-                        copy={props.copy}
-                        onChange={props.onChange}
-                        mode={props.mode}
-                        disabled={true}
-                    />
-                    <button className={`map-switcher-inline ${housing}-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={housingSwitchOnClick}>
-                        {(housing === 'enabled')? 'Click to hide Housing Zones' : 'Click to see Housing Zones mapped'}
-                    </button>
-                    <LogicalDataEntry
-                        slug='planning_enterprise_zone'
-                        title={dataFields.planning_enterprise_zone.title}
-                        tooltip={dataFields.planning_enterprise_zone.tooltip}
-                        value={props.building.planning_enterprise_zone}
-                        copy={props.copy}
-                        onChange={props.onChange}
-                        mode={props.mode}
-                        disabled={true}
-                    />
-                    <button className={`map-switcher-inline ${creative}-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={creativeSwitchOnClick}>
-                        {(creative === 'enabled')? 'Click to hide Creative Enterprise Zones' : 'Click to see Creative Enterprise Zones'}
-                    </button>
-                    <LogicalDataEntry
-                        slug='planning_protected_vista'
-                        title={dataFields.planning_protected_vista.title}
-                        tooltip={dataFields.planning_protected_vista.tooltip}
-                        value={props.building.planning_protected_vista}
-                        copy={props.copy}
-                        onChange={props.onChange}
-                        mode={props.mode}
-                        disabled={true}
-                    />
-                    <button className={`map-switcher-inline ${vista}-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={vistaSwitchOnClick}>
-                        {(vista === 'enabled')? 'Click to hide Protected Vistas' : 'Click to see Protected Vistas'}
-                    </button>
-                    {/*
-                        <DataEntry
-                        title={dataFields.planning_glher_url.title}
-                        slug="planning_glher_url"
-                        value={props.building.planning_glher_url}
-                        mode={props.mode}
-                        copy={props.copy}
-                        onChange={props.onChange}
-                        isUrl={true}
-                        placeholder="Please add relevant link here"
-                        />
-                    <Verification
-                        slug="planning_glher_url"
-                        allow_verify={props.user !== undefined && props.building.planning_glher_url !== null && !props.edited}
-                        onVerify={props.onVerify}
-                        user_verified={props.user_verified.hasOwnProperty("planning_glher_url")}
-                        user_verified_as={props.user_verified.planning_glher_url}
-                        verified_count={props.building.verified.planning_glher_url}
-                        />
-                    */}
-                </DataEntryGroup>    
-                <DataEntryGroup name="Land ownership type" collapsed={subcat==null || subcat!="8"}>
-                    <InfoBox>
-                        This section is designed to provide information on land parcels and their ownership type. Can you help us collect this information?
-                    </InfoBox>
-                    <SelectDataEntry
-                        slug='community_public_ownership'
-                        title={dataFields.community_public_ownership.title}
-                        value={props.building.community_public_ownership}
-                        options={dataFields.community_public_ownership.items}
-                        tooltip={dataFields.community_public_ownership.tooltip}
-                        onChange={props.onChange}
-                        mode={props.mode}
-                        copy={props.copy}
-                    />
-                    <Verification
-                        slug="community_public_ownership"
-                        allow_verify={props.user !== undefined && props.building.community_public_ownership !== null && !props.edited}
-                        onVerify={props.onVerify}
-                        user_verified={props.user_verified.hasOwnProperty("community_public_ownership")}
-                        user_verified_as={props.user_verified.community_public_ownership}
-                        verified_count={props.building.verified.community_public_ownership}
-                    />
-                    <SelectDataEntry
-                        title={dataFields.community_public_ownership_source_type.title}
-                        slug="community_public_ownership_source_type"
-                        value={props.building.community_public_ownership_source_type}
-                        mode={props.mode}
-                        copy={props.copy}
-                        onChange={props.onChange}
-                        tooltip={dataFields.community_public_ownership_source_type.tooltip}
-                        options={dataFields.community_public_ownership_source_type.items}
-                        placeholder={dataFields.community_public_ownership_source_type.example}
-                    />
-                    {(props.building.community_public_ownership_source_type == dataFields.community_public_ownership_source_type.items[0] ||
-                        props.building.community_public_ownership_source_type == dataFields.community_public_ownership_source_type.items[1] ||
-                        props.building.community_public_ownership_source_type == null) ? <></> :
-                        <>
-                            <MultiDataEntry
-                                slug='community_public_ownership_sources'
-                                title={dataFields.community_public_ownership_sources.title}
-                                tooltip={dataFields.community_public_ownership_sources.tooltip}
-                                isUrl={true}
-                                placeholder={'https://...'}
-                                editableEntries={true}
-                                value={props.building.community_public_ownership_sources}
-                                onChange={props.onChange}
-                                mode={props.mode}
-                                copy={props.copy}
-                            />
-                        </>
-                    }
-                    <hr/>
-                    <DataEntry
-                        title={dataFields.size_parcel_geometry.title}
-                        slug="size_parcel_geometry"
-                        value={props.building.size_parcel_geometry}
-                        mode={props.mode}
-                        onChange={props.onChange}
-                        tooltip={dataFields.size_parcel_geometry.tooltip}
-                        placeholder="https://..."
-                        isUrl={true}
-                    />
-                    <button className={`map-switcher-inline ${parcel}-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={parcelSwitchOnClick}>
-                        {(parcel === 'enabled')? 'Click to hide sample land parcel data' : 'Click to show sample land parcel data'}
-                    </button>
-                </DataEntryGroup>
+                <InfoBox type='warning'>
+                    This section is under development in collaboration with the historic environment sector.
+                    Please let us know your suggestions on the <a href="https://discuss.colouring.london/t/dynamics-category-discussion/107">discussion forum</a>! (external link - save your edits first)
+                </InfoBox>
             </DataEntryGroup>
-            <DataEntryGroup name="Heritage assets and building protection" collapsed={subcat==null || subcat!="7"}>
-                <InfoBox>
-                Help us produce the most accurate map possible for London's designated/protected buildings. Please add data if missing or click "Verify" where entries are correct.
+            <DataEntryGroup name="Survival tracking" collapsed={subcat==null || subcat!="5"}>
+                <div className={`alert alert-dark`} role="alert" style={{ fontSize: 13, backgroundColor: "#f6f8f9" }}>
+                    <i>
+                        Can you help us create a map that shows how many buildings in this area have survived since the 1890s? 
+                        Choose a colour to indicate whether the building has survived.
+                    </i>
+                </div>
+                {(historicMap === "enabled") ? 
+                    <button className={`map-switcher-inline enabled-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={switchToAgeMapStyle}>
+                        Click to hide the 1890s OS historical map.
+                    </button>
+                :
+                    <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={switchToSurvivalMapStyle}>
+                        Click to show the 1890s OS historical map.
+                    </button>
+                }
+                {(historicData === "enabled") ? 
+                    <button className={`map-switcher-inline enabled-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={switchToAgeMapStyle}>
+                        Click to hide the 1890s OS historical map with modern footprints.
+                    </button>
+                :
+                    <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={switchToSurvivalDataStyle}>
+                        Click to show the 1890s OS historical map with modern footprints.
+                    </button>
+                }
+                <SelectDataEntry
+                    title={dataFields.survival_status.title}
+                    slug="survival_status"
+                    value={props.building.survival_status}
+                    tooltip={dataFields.survival_status.tooltip}
+                    options={dataFields.survival_status.items}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                />
+                <SelectDataEntry
+                    title={dataFields.survival_source.title}
+                    slug="survival_source"
+                    value={props.building.survival_source}
+                    mode={props.mode}
+                    copy={props.copy}
+                    onChange={props.onChange}
+                    tooltip={dataFields.survival_source.tooltip}
+                    placeholder={dataFields.survival_source.example}
+                    options={dataFields.survival_source.items}
+                />
+                {(props.building.survival_source == dataFields.survival_source.items[0] ||
+                    props.building.survival_source == dataFields.survival_source.items[1] ||
+                    props.building.survival_source == null) ? <></> :
+                    <><MultiDataEntry
+                        title={dataFields.survival_source_links.title}
+                        slug="survival_source_links"
+                        value={props.building.survival_source_links}
+                        mode={props.mode}
+                        copy={props.copy}
+                        onChange={props.onChange}
+                        tooltip={dataFields.survival_source_links.tooltip}
+                        placeholder="https://..."
+                        editableEntries={true}
+                        isUrl={true}
+                        />
+                    </>
+                }
+            </DataEntryGroup>
+            <DataEntryGroup name="Historical map data options" collapsed={subcat==null || subcat!="6"}>
+                <InfoBox type='warning'>
+                    This section is under development
                 </InfoBox>
                 <div className={`alert alert-dark`} role="alert" style={{ fontSize: 13, backgroundColor: "#f6f8f9" }}>
-                    <i><div><u>Disclaimer</u>:  Data for designated heritage assets has been accessed from the <a href="https://historicengland.org.uk/listing/the-list/">National Heritage List for England</a>. Source information for Conservation Area data can be accessed <a href="http://www.bedfordpark.net/leo/planning/">here</a>. Please note all data should be double checked against official sources where used for planning purposes'.</div></i>
+                    <i>
+                        This section provides links to open digitised historical maps/mapping data that we are using in the Colouring Cities platform.
+                    </i>
                 </div>
-                {props.mapColourScale != "planning_combined" ?
-                    <button className={`map-switcher-inline disabled-state btn btn-outline btn-outline-dark key-button`} onClick={switchToBuildingProtectionMapStyle}>
-                        {'Click to see individual protected buildings mapped'}
-                    </button>
-                    :
-                    <></>
-                }
-                <button className={`map-switcher-inline ${conservation}-state btn btn-outline btn-outline-dark ${darkLightTheme}`} onClick={conservationSwitchOnClick}>
-                    {(conservation === 'enabled')? 'Click to hide Conservation Areas' : 'Click to see Conservation Areas'}
-                </button>
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_heritage_at_risk'
-                    title={dataFields.planning_heritage_at_risk.title}
-                    tooltip={dataFields.planning_heritage_at_risk.tooltip}
-                    value={props.building.planning_heritage_at_risk}
+                <MultiDataEntry
+                    title={dataFields.age_historical_raster_map_links.title}
+                    slug="age_historical_raster_map_links"
+                    value={props.building.age_historical_raster_map_links}
+                    mode={props.mode}
                     copy={props.copy}
                     onChange={props.onChange}
+                    tooltip={dataFields.age_historical_raster_map_links.tooltip}
+                    placeholder="https://..."
+                    editableEntries={true}
+                    isUrl={true}
+                />
+                <MultiDataEntry
+                    title={dataFields.age_historical_vectorised_footprint_links.title}
+                    slug="age_historical_vectorised_footprint_links"
+                    value={props.building.age_historical_vectorised_footprint_links}
                     mode={props.mode}
-                />
-                <Verification
-                    slug="planning_heritage_at_risk"
-                    allow_verify={props.user !== undefined && props.building.planning_heritage_at_risk !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_heritage_at_risk")}
-                    user_verified_as={props.user_verified.planning_heritage_at_risk}
-                    verified_count={props.building.verified.planning_heritage_at_risk}
-                />
-                {(props.building.planning_heritage_at_risk == null || props.building.planning_heritage_at_risk == false) ? <></> :
-                    <>
-                        <DataEntry
-                            title={dataFields.planning_heritage_at_risk_url.title}
-                            slug="planning_heritage_at_risk_url"
-                            value={props.building.planning_heritage_at_risk_url}
-                            tooltip={dataFields.planning_heritage_at_risk_url.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            placeholder="Please add relevant link here"
-                            isUrl={true}
-                            />
-                        <Verification
-                            slug="planning_heritage_at_risk_url"
-                            allow_verify={props.user !== undefined && props.building.planning_heritage_at_risk_url !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_heritage_at_risk_url")}
-                            user_verified_as={props.user_verified.planning_heritage_at_risk_url}
-                            verified_count={props.building.verified.planning_heritage_at_risk_url}
-                            />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_world_heritage_site'
-                    title={dataFields.planning_world_heritage_site.title}
-                    tooltip={dataFields.planning_world_heritage_site.tooltip}
-                    value={props.building.planning_world_heritage_site}
                     copy={props.copy}
                     onChange={props.onChange}
-                    mode={props.mode}
+                    tooltip={dataFields.age_historical_vectorised_footprint_links.tooltip}
+                    placeholder="https://..."
+                    editableEntries={true}
+                    isUrl={true}
                 />
-                <Verification
-                    slug="planning_world_heritage_site"
-                    allow_verify={props.user !== undefined && props.building.planning_world_heritage_site !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_world_heritage_site")}
-                    user_verified_as={props.user_verified.planning_world_heritage_site}
-                    verified_count={props.building.verified.planning_world_heritage_site}
-                />
-                {(props.building.planning_world_heritage_site == null || props.building.planning_world_heritage_site == false) ? <></> :
-                    <>
-                        <NumericDataEntryWithFormattedLink
-                            title={dataFields.planning_world_list_id.title}
-                            slug="planning_world_list_id"
-                            value={props.building.planning_world_list_id}
-                            tooltip={dataFields.planning_world_list_id.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            placeholder="add ID here"
-                            linkTargetFunction={(id: String) => { return "https://whc.unesco.org/en/list/" + id } }
-                            linkDescriptionFunction={(id: String) => { return "UNESCO Link" } }
-                        />
-                        <Verification
-                            slug="planning_world_list_id"
-                            allow_verify={props.user !== undefined && props.building.planning_world_list_id !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_world_list_id")}
-                            user_verified_as={props.user_verified.planning_world_list_id}
-                            verified_count={props.building.verified.planning_world_list_id}
-                        />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_listed'
-                    title={dataFields.planning_listed.title}
-                    tooltip={dataFields.planning_listed.tooltip}
-                    value={props.building.planning_listed}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_listed"
-                    allow_verify={props.user !== undefined && props.building.planning_listed !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_listed")}
-                    user_verified_as={props.user_verified.planning_listed}
-                    verified_count={props.building.verified.planning_listed}
-                />
-                {(props.building.planning_listed == null || props.building.planning_listed == false) ? <></> :
-                    <>
-                        <SelectDataEntry
-                            title={dataFields.planning_list_grade.title}
-                            slug="planning_list_grade"
-                            value={props.building.planning_list_grade}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            tooltip={dataFields.planning_list_grade.tooltip}
-                            options={dataFields.planning_list_grade.items}
-                        />
-                        <NumericDataEntryWithFormattedLink
-                            title={dataFields.planning_list_id.title}
-                            slug="planning_list_id"
-                            value={props.building.planning_list_id}
-                            tooltip={dataFields.planning_list_id.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            placeholder="add ID here"
-                            linkTargetFunction={(id: String) => { return "https://historicengland.org.uk/listing/the-list/list-entry/" + id } }
-                            linkDescriptionFunction={(id: String) => { return "Historic England Link" } }
-                        />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_local_list'
-                    title={dataFields.planning_local_list.title}
-                    tooltip={dataFields.planning_local_list.tooltip}
-                    value={props.building.planning_local_list}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_local_list"
-                    allow_verify={props.user !== undefined && props.building.planning_local_list !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_local_list")}
-                    user_verified_as={props.user_verified.planning_local_list}
-                    verified_count={props.building.verified.planning_local_list}
-                />
-                {(props.building.planning_local_list == null || props.building.planning_local_list == false) ? <></> :
-                    <>
-                        <DataEntry
-                            title={dataFields.planning_local_list_url.title}
-                            slug="planning_local_list_url"
-                            value={props.building.planning_local_list_url}
-                            tooltip={dataFields.planning_local_list_url.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            isUrl={true}
-                            placeholder="Please add relevant link here"
-                            />
-                        <Verification
-                            slug="planning_local_list_url"
-                            allow_verify={props.user !== undefined && props.building.planning_local_list_url !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_local_list_url")}
-                            user_verified_as={props.user_verified.planning_local_list_url}
-                            verified_count={props.building.verified.planning_local_list_url}
-                            />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_in_conservation_area'
-                    title={dataFields.planning_in_conservation_area.title}
-                    tooltip={dataFields.planning_in_conservation_area.tooltip}
-                    value={props.building.planning_in_conservation_area}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_in_conservation_area"
-                    allow_verify={props.user !== undefined && props.building.planning_in_conservation_area !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_in_conservation_area")}
-                    user_verified_as={props.user_verified.planning_in_conservation_area}
-                    verified_count={props.building.verified.planning_in_conservation_area}
-                />
-                {(props.building.planning_in_conservation_area == null || props.building.planning_in_conservation_area == false) ? <></> :
-                    <>
-                        <DataEntry
-                            title={dataFields.planning_in_conservation_area_url.title}
-                            slug="planning_in_conservation_area_url"
-                            value={props.building.planning_in_conservation_area_url}
-                            tooltip={dataFields.planning_in_conservation_area_url.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            isUrl={true}
-                            placeholder="Please add CA appraisal link here"
-                            />
-                        {props.building.planning_in_conservation_area_url === "" ? "Our CA map records this building as not being within a CA. To help us verify this, please click ‘verify’ or, if info is incorrect, please add the local authority’s CA appraisal link." : "" }
-                        {props.building.planning_in_conservation_area_url === "identified as listed: please replace with links" ? "Our CA map records this building as being within a CA. To help us verify this information please add the local authority’s CA appraisal link and then click ‘verify’." : "" }
-                        <Verification
-                            slug="planning_in_conservation_area_url"
-                            allow_verify={props.user !== undefined && props.building.planning_in_conservation_area_url !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_in_conservation_area_url")}
-                            user_verified_as={props.user_verified.planning_in_conservation_area_url}
-                            verified_count={props.building.verified.planning_in_conservation_area_url}
-                        />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_in_apa'
-                    title={dataFields.planning_in_apa.title}
-                    tooltip={dataFields.planning_in_apa.tooltip}
-                    value={props.building.planning_in_apa}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_in_apa"
-                    allow_verify={props.user !== undefined && props.building.planning_in_apa !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_in_apa")}
-                    user_verified_as={props.user_verified.planning_in_apa}
-                    verified_count={props.building.verified.planning_in_apa}
-                />
-                {(props.building.planning_in_apa == null || props.building.planning_in_apa == false) ? <></> :
-                    <>
-                        <DataEntry
-                            title={dataFields.planning_in_apa_url.title}
-                            slug="planning_in_apa_url"
-                            value={props.building.planning_in_apa_url}
-                            tooltip={dataFields.planning_in_apa_url.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            isUrl={true}
-                            placeholder="Please add relevant link here"
-                            />
-                        <Verification
-                            slug="planning_in_apa_url"
-                            allow_verify={props.user !== undefined && props.building.planning_in_apa_url !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_in_apa_url")}
-                            user_verified_as={props.user_verified.planning_in_apa_url}
-                            verified_count={props.building.verified.planning_in_apa_url}
-                            />
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_scientific_interest'
-                    title={dataFields.planning_scientific_interest.title}
-                    tooltip={dataFields.planning_scientific_interest.tooltip}
-                    value={props.building.planning_scientific_interest}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_scientific_interest"
-                    allow_verify={props.user !== undefined && props.building.planning_scientific_interest !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_scientific_interest")}
-                    user_verified_as={props.user_verified.planning_scientific_interest}
-                    verified_count={props.building.verified.planning_scientific_interest}
-                />
-                {(props.building.planning_scientific_interest == null || props.building.planning_scientific_interest == false) ? <></> :
-                    <>
-                        {/*<SelectDataEntry
-                            title={dataFields.planning_scientific_interest_source_type.title}
-                            slug="planning_scientific_interest_source_type"
-                            value={props.building.planning_scientific_interest_source_type}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            tooltip={dataFields.planning_scientific_interest_source_type.tooltip}
-                            options={dataFields.planning_scientific_interest_source_type.items}
-                            placeholder={dataFields.planning_scientific_interest_source_type.example}
-                        />
-                        {(props.building.planning_scientific_interest_source_type == dataFields.planning_scientific_interest_source_type.items[0] ||
-                            props.building.planning_scientific_interest_source_type == dataFields.planning_scientific_interest_source_type.items[1] ||
-                            props.building.planning_scientific_interest_source_type == null) ? <></> :
-                            <>
-                        */}
-                                <MultiDataEntry
-                                    title={dataFields.planning_scientific_interest_source_links.title}
-                                    slug="planning_scientific_interest_source_links"
-                                    value={props.building.planning_scientific_interest_source_links}
-                                    mode={props.mode}
-                                    copy={props.copy}
-                                    onChange={props.onChange}
-                                    tooltip={dataFields.planning_scientific_interest_source_links.tooltip}
-                                    placeholder="https://..."
-                                    editableEntries={true}
-                                    isUrl={true}
-                                />
-                            {/* </>
-                        }*/}
-                    </>
-                }
-                <hr/>
-                <LogicalDataEntry
-                    slug='planning_historic_area_assessment'
-                    title={dataFields.planning_historic_area_assessment.title}
-                    tooltip={dataFields.planning_historic_area_assessment.tooltip}
-                    value={props.building.planning_historic_area_assessment}
-                    copy={props.copy}
-                    onChange={props.onChange}
-                    mode={props.mode}
-                />
-                <Verification
-                    slug="planning_historic_area_assessment"
-                    allow_verify={props.user !== undefined && props.building.planning_historic_area_assessment !== null && !props.edited}
-                    onVerify={props.onVerify}
-                    user_verified={props.user_verified.hasOwnProperty("planning_historic_area_assessment")}
-                    user_verified_as={props.user_verified.planning_historic_area_assessment}
-                    verified_count={props.building.verified.planning_historic_area_assessment}
-                />
-                {(props.building.planning_historic_area_assessment == null || props.building.planning_historic_area_assessment == false) ? <></> :
-                    <>
-                        <DataEntry
-                            title={dataFields.planning_historic_area_assessment_url.title}
-                            slug="planning_historic_area_assessment_url"
-                            value={props.building.planning_historic_area_assessment_url}
-                            tooltip={dataFields.planning_historic_area_assessment_url.tooltip}
-                            mode={props.mode}
-                            copy={props.copy}
-                            onChange={props.onChange}
-                            isUrl={true}
-                            placeholder="Please add relevant link here"
-                            />
-                        <Verification
-                            slug="planning_historic_area_assessment_url"
-                            allow_verify={props.user !== undefined && props.building.planning_historic_area_assessment_url !== null && !props.edited}
-                            onVerify={props.onVerify}
-                            user_verified={props.user_verified.hasOwnProperty("planning_historic_area_assessment_url")}
-                            user_verified_as={props.user_verified.planning_historic_area_assessment_url}
-                            verified_count={props.building.verified.planning_historic_area_assessment_url}
-                            />
-                    </>
-                }
-            </DataEntryGroup>  
+            </DataEntryGroup>
         </Fragment>
     );
 };
